pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PyStation"
version = "0.0.1"
description = "Playstation Store price crawler"
readme = "README.md"
authors = [
    { name="Paavo Westerlund"}
]
license = "MIT"
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    'Programming Language :: Python :: 3',
]
keywords = ["Playstation", "Store"]
dependencies = [
    "beautifulsoup4>=4.11.1",
    "Pillow>=9.2.0",
    "requests>=2.28.1",
    "tabulate>=0.8.10",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["black", "pytest"]

[project.urls]
"Homepage" = "https://github.com/paavoto7/PyStation"

[project.scripts]
pystation = "pystation.cli.cli:main"